extract SharedVariableAccessMethodTransformer into separate class
